Move UI and Windows to folders
