+++ trview.app/UI/CameraControls.cpp
