+++ trview.app/UI/CameraPosition.cpp
