+++ trview.app/UI/CameraPosition.h
