+++ trview.app/UI/Flipmaps.cpp
