+++ trview.app/UI/Flipmaps.h
