+++ trview.app/UI/LevelInfo.cpp
@@ -5,7 +5,7 @@
 #include <trview.ui/Label.h>
 #include <trview.ui/StackPanel.h>
 #include <trview.ui/Button.h>
-#include "ITextureStorage.h"
+#include <trview.app/ITextureStorage.h>
 #include <trview.common/Strings.h>
 
 namespace trview
