+++ trview.app/UI/LevelInfo.h
