+++ trview.app/UI/RoomNavigator.h
