+++ trview.app/UI/SettingsWindow.cpp
