+++ trview.app/UI/SettingsWindow.h
