+++ trview.app/Windows/CollapsiblePanel.cpp
