+++ trview.app/Windows/ItemsWindow.cpp
