+++ trview.app/Windows/ItemsWindowManager.cpp
@@ -1,5 +1,5 @@
 #include "ItemsWindowManager.h"
-#include "WindowIDs.h"
+#include <trview.app/WindowIDs.h>
 #include <trview.ui.render/Renderer.h>
 #include <trview.graphics/DeviceWindow.h>
