+++ trview.app/Windows/ItemsWindowManager.h
