+++ trview.app/Windows/RouteWindowManager.h
