+++ trview.app/Windows/TriggersWindow.cpp
