+++ trview.app/Windows/TriggersWindow.h
@@ -4,8 +4,8 @@
 #pragma once
 
 #include <trview.ui/Listbox.h>
-#include "Item.h"
-#include "Trigger.h"
+#include <trview.app/Item.h>
+#include <trview.app/Trigger.h>
 #include "CollapsiblePanel.h"
 
 namespace trview
