+++ trview.app/Windows/TriggersWindowManager.h
