+++ trview.app/trview.app.vcxproj
@@ -21,109 +21,108 @@
   <ItemGroup>
     <ClCompile Include="AlternateGroupToggler.cpp" />
     <ClCompile Include="Camera\Camera.cpp" />
-    <ClCompile Include="Camera\CameraControls.cpp" />
     <ClCompile Include="Camera\CameraMode.cpp" />
-    <ClCompile Include="Camera\CameraPosition.cpp" />
     <ClCompile Include="Camera\FreeCamera.cpp" />
     <ClCompile Include="Camera\ICamera.cpp" />
     <ClCompile Include="Camera\OrbitCamera.cpp" />
-    <ClCompile Include="CollapsiblePanel.cpp" />
     <ClCompile Include="Compass.cpp" />
     <ClCompile Include="ContextMenu.cpp" />
     <ClCompile Include="DirectoryListing.cpp" />
     <ClCompile Include="FileDropper.cpp" />
-    <ClCompile Include="Flipmaps.cpp" />
     <ClCompile Include="Geometry\IRenderable.cpp" />
     <ClCompile Include="Geometry\Mesh.cpp" />
     <ClCompile Include="Geometry\TransparencyBuffer.cpp" />
     <ClCompile Include="Geometry\TransparentTriangle.cpp" />
-    <ClCompile Include="GoTo.cpp" />
     <ClCompile Include="ILevelTextureStorage.cpp" />
     <ClCompile Include="Item.cpp" />
-    <ClCompile Include="ItemsWindow.cpp" />
-    <ClCompile Include="ItemsWindowManager.cpp" />
     <ClCompile Include="ITextureStorage.cpp" />
-    <ClCompile Include="LevelInfo.cpp" />
     <ClCompile Include="LevelSwitcher.cpp" />
     <ClCompile Include="Measure.cpp" />
-    <ClCompile Include="Neighbours.cpp" />
     <ClCompile Include="Picking.cpp" />
     <ClCompile Include="PickResult.cpp" />
     <ClCompile Include="RecentFiles.cpp" />
-    <ClCompile Include="RoomNavigator.cpp" />
     <ClCompile Include="Route.cpp" />
-    <ClCompile Include="RouteWindow.cpp" />
-    <ClCompile Include="RouteWindowManager.cpp" />
     <ClCompile Include="Sector.cpp" />
     <ClCompile Include="SectorHighlight.cpp" />
     <ClCompile Include="SelectionRenderer.cpp" />
-    <ClCompile Include="SettingsWindow.cpp" />
     <ClCompile Include="Toolbar.cpp" />
     <ClCompile Include="Trigger.cpp" />
-    <ClCompile Include="TriggersWindow.cpp" />
-    <ClCompile Include="TriggersWindowManager.cpp" />
+    <ClCompile Include="UI\CameraControls.cpp" />
+    <ClCompile Include="UI\CameraPosition.cpp" />
+    <ClCompile Include="UI\Flipmaps.cpp" />
+    <ClCompile Include="UI\GoTo.cpp" />
+    <ClCompile Include="UI\LevelInfo.cpp" />
+    <ClCompile Include="UI\Neighbours.cpp" />
+    <ClCompile Include="UI\RoomNavigator.cpp" />
+    <ClCompile Include="UI\SettingsWindow.cpp" />
+    <ClCompile Include="UI\ViewerUI.cpp" />
     <ClCompile Include="UserSettings.cpp" />
-    <ClCompile Include="ViewerUI.cpp" />
     <ClCompile Include="ViewMenu.cpp" />
     <ClCompile Include="Waypoint.cpp" />
     <ClCompile Include="WindowResizer.cpp" />
+    <ClCompile Include="Windows\CollapsiblePanel.cpp" />
+    <ClCompile Include="Windows\ItemsWindow.cpp" />
+    <ClCompile Include="Windows\ItemsWindowManager.cpp" />
+    <ClCompile Include="Windows\RouteWindow.cpp" />
+    <ClCompile Include="Windows\RouteWindowManager.cpp" />
+    <ClCompile Include="Windows\TriggersWindow.cpp" />
+    <ClCompile Include="Windows\TriggersWindowManager.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="AlternateGroupToggler.h" />
-    <ClInclude Include="CameraControls.h" />
     <ClInclude Include="Camera\Camera.h" />
-    <ClInclude Include="Camera\CameraControls.h" />
     <ClInclude Include="Camera\CameraMode.h" />
-    <ClInclude Include="Camera\CameraPosition.h" />
     <ClInclude Include="Camera\FreeCamera.h" />
     <ClInclude Include="Camera\ICamera.h" />
     <ClInclude Include="Camera\OrbitCamera.h" />
-    <ClInclude Include="CollapsiblePanel.h" />
     <ClInclude Include="Compass.h" />
     <ClInclude Include="ContextMenu.h" />
     <ClInclude Include="DirectoryListing.h" />
     <ClInclude Include="FileDropper.h" />
-    <ClInclude Include="Flipmaps.h" />
     <ClInclude Include="Geometry\IRenderable.h" />
     <ClInclude Include="Geometry\Mesh.h" />
     <ClInclude Include="Geometry\MeshVertex.h" />
     <ClInclude Include="Geometry\TransparencyBuffer.h" />
     <ClInclude Include="Geometry\TransparentTriangle.h" />
     <ClInclude Include="Geometry\Triangle.h" />
-    <ClInclude Include="GoTo.h" />
     <ClInclude Include="ILevelTextureStorage.h" />
     <ClInclude Include="Item.h" />
-    <ClInclude Include="ItemsWindow.h" />
-    <ClInclude Include="ItemsWindowManager.h" />
     <ClInclude Include="ITextureStorage.h" />
-    <ClInclude Include="LevelInfo.h" />
     <ClInclude Include="LevelSwitcher.h" />
     <ClInclude Include="Measure.h" />
-    <ClInclude Include="Neighbours.h" />
     <ClInclude Include="PickInfo.h" />
     <ClInclude Include="Picking.h" />
     <ClInclude Include="PickResult.h" />
     <ClInclude Include="RecentFiles.h" />
     <ClInclude Include="RoomInfo.h" />
-    <ClInclude Include="RoomNavigator.h" />
     <ClInclude Include="Route.h" />
-    <ClInclude Include="RouteWindow.h" />
-    <ClInclude Include="RouteWindowManager.h" />
     <ClInclude Include="Sector.h" />
     <ClInclude Include="SectorHighlight.h" />
     <ClInclude Include="SelectionRenderer.h" />
-    <ClInclude Include="SettingsWindow.h" />
     <ClInclude Include="Toolbar.h" />
     <ClInclude Include="Trigger.h" />
-    <ClInclude Include="TriggersWindow.h" />
-    <ClInclude Include="TriggersWindowManager.h" />
     <ClInclude Include="Types.h" />
+    <ClInclude Include="UI\CameraControls.h" />
+    <ClInclude Include="UI\CameraPosition.h" />
+    <ClInclude Include="UI\Flipmaps.h" />
+    <ClInclude Include="UI\GoTo.h" />
+    <ClInclude Include="UI\LevelInfo.h" />
+    <ClInclude Include="UI\Neighbours.h" />
+    <ClInclude Include="UI\RoomNavigator.h" />
+    <ClInclude Include="UI\SettingsWindow.h" />
+    <ClInclude Include="UI\ViewerUI.h" />
     <ClInclude Include="UserSettings.h" />
-    <ClInclude Include="ViewerUI.h" />
     <ClInclude Include="ViewMenu.h" />
     <ClInclude Include="Waypoint.h" />
     <ClInclude Include="WindowIDs.h" />
     <ClInclude Include="WindowResizer.h" />
+    <ClInclude Include="Windows\CollapsiblePanel.h" />
+    <ClInclude Include="Windows\ItemsWindow.h" />
+    <ClInclude Include="Windows\ItemsWindowManager.h" />
+    <ClInclude Include="Windows\RouteWindow.h" />
+    <ClInclude Include="Windows\RouteWindowManager.h" />
+    <ClInclude Include="Windows\TriggersWindow.h" />
+    <ClInclude Include="Windows\TriggersWindowManager.h" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\trview.common\trview.common.vcxproj">
