+++ trview/Viewer.h
@@ -28,17 +28,17 @@
 #include <trview.app/WindowResizer.h>
 #include <trview.app/RecentFiles.h>
 #include <trview.app/FileDropper.h>
-#include <trview.app/ItemsWindowManager.h>
-#include <trview.app/TriggersWindowManager.h>
+#include <trview.app/Windows/ItemsWindowManager.h>
+#include <trview.app/Windows/TriggersWindowManager.h>
 #include <trview.app/Measure.h>
 #include <trview.app/Compass.h>
 #include <trview.app/AlternateGroupToggler.h>
 #include <trview.app/Route.h>
-#include <trview.app/RouteWindowManager.h>
+#include <trview.app/Windows/RouteWindowManager.h>
 #include <trview.app/ViewMenu.h>
 #include <trview.app/Picking.h>
 #include <trview.app/SectorHighlight.h>
-#include <trview.app/ViewerUI.h>
+#include <trview.app/UI/ViewerUI.h>
 
 namespace trview
 {
